Some details

+++ Analyzer/src/main/java/com/needine/spark/Tables.scala
@@ -6,7 +6,7 @@ object Tables {
   
   //case class Packet(time: Timestamp, origen:Double, destiny: Double, bytes: Double)extends Serializable
   
-  case class Packet(time: Long, origen:Double, destiny: Double, bytes: Double)extends Serializable
+  case class Packet(time: Long, origen:Long, destiny: Long, bytes: Double)extends Serializable
   
   
   case class Origin_By_IP_TCP(ip: String, ref: Double)extends Serializable
